fix: eliminate circular import between in_memory.py and in_memory_factories.py
Merge the 2 remaining factory functions (get_event_store,
get_event_publisher) into in_memory.py and delete in_memory_factories.py.
The factories were originally split out when there were 9 repo factories —
now there are only 2, so a separate file creates unnecessary circular
dependency risk.

Addresses Copilot review comment on #616.
